@@ -6,13 +6,13 @@
   "scripts": {
     "webpack": "webpack",
     "test": "jest && lab -v -l",
-    "lint": "eslint src/react_components/main.js",
+    "lint": "eslint src/components/main.js",
     "watch": "npm-watch",
     "start": "node walmart.js",
     "postinstall": "mkdir -p public/js && cp node_modules/bootstrap/dist/js/bootstrap.min.js node_modules/jquery/dist/jquery.min.* public/js && cp node_modules/bootstrap/dist/css/bootstrap*.min.css public/css && npm run webpack"
   },
   "engines": {
-    "node": "^0.12"
+    "node": "6"
   },
   "repository": {
     "type": "git",
@@ -23,7 +23,8 @@
     "hapi",
     "node",
     "Facebook",
-    "Walmart"
+    "Walmart",
+    "mullet"
   ],
   "author": "Linda Nichols",
   "license": "MIT",
@@ -32,44 +33,36 @@
   },
   "homepage": "http://www.mullet.io",
   "dependencies": {
-    "babel": "^5.8.23",
+    "babel": "^6.23.0",
     "babel-core": "^6.4.0",
-    "babel-jest": "^5.3.0",
+    "babel-jest": "^19.0.0",
     "babel-loader": "^6.2.1",
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "blipp": "^2.3.0",
     "body-parser": "^1.14.1",
     "bootstrap": "^3.3.5",
-    "h2o2": "^4.0.2",
-    "hapi": "^9.3.1",
-    "inert": "^3.2.0",
-    "jquery": "^2.1.4",
-    "npm-watch": "^0.0.1",
-    "react": "^0.13.3",
-    "vision": "^3.0.0",
-    "webpack": "^1.12.10"
+    "h2o2": "^5.4.0",
+    "hapi": "^16.1.0",
+    "inert": "^4.1.0",
+    "jest": "^19.0.2",
+    "jquery": "^3.2.0",
+    "npm-watch": "^0.1.8",
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-scripts": "^0.9.5",
+    "vision": "^4.1.1",
+    "webpack": "^2.2.1"
   },
   "devDependencies": {
-    "eslint": "^1.6.0",
-    "eslint-plugin-react": "^3.5.1",
-    "jest-cli": "^0.4.19",
-    "code": "^1.5.0",
-    "lab": "^5.15.2"
-  },
-  "jest": {
-    "unmockedModulePathPatterns": [
-      "react"
-    ],
-    "scriptPreprocessor": "<rootDir>/node_modules/babel-jest",
-    "testFileExtensions": [
-      "es6",
-      "js"
-    ],
-    "moduleFileExtensions": [
-      "js",
-      "json",
-      "es6"
-    ]
+    "code": "^4.0.0",
+    "eslint": "^3.18.0",
+    "eslint-config-eta": "^0.0.7",
+    "eslint-plugin-babel": "^4.1.1",
+    "eslint-plugin-react": "^6.10.0",
+    "jest-cli": "^19.0.2",
+    "lab": "^13.0.1",
+    "enzyme": "^2.7.1",
+    "react-test-renderer": "^15.4.2"
   }
 }
