@@ -1,13 +1,13 @@
 'use strict';
 
-var Code = require('code'); // Hapi assertion library
-var Lab = require('lab');
-var lab = exports.lab = Lab.script();
-var server = require('../walmart');
+const Code = require('code'); // assertion library for Lab
+const Lab = require('lab');
+const lab = exports.lab = Lab.script();
+const server = require('../walmart');
 
-lab.experiment('General Endpoint Tests', function() {
+lab.experiment('General Endpoint Tests', () => {
 
-  lab.test('GET / (default test)', function (done) {
+  lab.test('GET / (default test)', (done) => {
     var options = {
         method: 'GET',
         url: '/'
@@ -18,23 +18,23 @@ lab.experiment('General Endpoint Tests', function() {
     });
   });
 
-  lab.test('GET /images/mullet_600.png', function (done) {
+  lab.test('GET /images/mullet_600.png', (done) => {
     var options = {
         method: 'GET',
         url: '/images/mullet_600.png'
     };
-    server.inject(options, function(response) {
+    server.inject(options, (response) => {
       Code.expect(response.statusCode).to.equal(200);
       done();
     });
   });
 
-  lab.test('GET /data', function (done) {
+  lab.test('GET /data', (done) => {
     var options = {
         method: 'GET',
         url: '/data'
     };
-    server.inject(options, function(response) {
+    server.inject(options, (response) => {
       Code.expect(response.statusCode).to.equal(200);
       Code.expect(response.result).to.be.an.object();
       Code.expect(response.result.message).to.equal('Welcome to Mullet!');
