@@ -1,8 +1,9 @@
-var path = require('path');
-var webpack = require('webpack');
+const path = require('path');
+const webpack = require('webpack');
+
 module.exports = {
   entry: [
-    './src/react_components/main.js'
+    './src/components/Main.js'
   ],
   output: {
     path: path.join(__dirname, 'public/js'),
@@ -10,7 +11,13 @@ module.exports = {
   },
   module: {
     loaders: [
-      { test: /.jsx?$/, loader: 'babel-loader', exclude: /node_modules/, query: { presets: ['es2015', 'react'] } }
+      { 
+        test: /.jsx?$/, 
+        loader: 'babel-loader', 
+        exclude: /node_modules/, 
+        query: { 
+          presets: ['es2015', 'react'] 
+        } }
     ]
   }
 };
